- int type fix (2)

+++ presto-cyoda/src/main/java/com/cyoda/presto/client/reporting/metaproviders/StaticReportTable.java
@@ -180,8 +180,8 @@ public enum StaticReportTable implements TableDefinition {
     public enum ApiCallStatsColumnDef implements ColumnDefinition {
         QUERY_ID(0, STRING),
         CALL_TIME(1, DATE),
-        CALL_MILLIS(2, LONG),
-        DURATION_MILLIS(3, INTEGER),
+        CALL_MILLIS(2, INTEGER),
+        DURATION_MILLIS(3, LONG),
         API_HANDLER(4, STRING),
         REQUEST_PARAMS(5, DataType.MAP, STRING, STRING),
         REQUEST_URL(6, STRING),
